Several structural changes and new modules for smoother use. Updated the way KIPET initializes models using simulation. This is large update and future updates will be more frequent.
